pyproject.toml
[project]
name = "smart-loadshedding-advisor"
version = "0.1.0"
description = "FastAPI + Streamlit app for the Smart Loadshedding Advisor"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "streamlit",
    "psycopg2",
    "langchain",
    "python-dotenv",
]
